+++ wasm/src/org.graalvm.wasm.test/src/org/graalvm/wasm/test/WasmFileSuite.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -306,18 +306,6 @@ public abstract class WasmFileSuite extends AbstractWasmSuite {
                     }
                 }
             });
-
-            final String signExtensionOpsProperty = "wasm.SignExtensionOps";
-            final String signExtensionOpsValue = testCase.options().getProperty(signExtensionOpsProperty);
-            if (signExtensionOpsValue != null) {
-                contextBuilder.option(signExtensionOpsProperty, signExtensionOpsValue);
-            }
-
-            final String multiValueProperty = "wasm.MultiValue";
-            final String multiValueValue = testCase.options().getProperty(multiValueProperty);
-            if (multiValueValue != null) {
-                contextBuilder.option(multiValueProperty, multiValueValue);
-            }
 
             final String envString = testCase.options().getProperty("env");
             if (envString != null) {