+++ wasm/src/org.graalvm.wasm/src/org/graalvm/wasm/api/InteropArray.java
@@ -45,7 +45,6 @@ import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.interop.InteropLibrary;
 import com.oracle.truffle.api.interop.InvalidArrayIndexException;
 import com.oracle.truffle.api.interop.TruffleObject;
-import com.oracle.truffle.api.interop.UnsupportedMessageException;
 import com.oracle.truffle.api.library.ExportLibrary;
 import com.oracle.truffle.api.library.ExportMessage;
 
@@ -62,6 +61,7 @@ public final class InteropArray implements TruffleObject {
     }
 
     @ExportMessage
+    @SuppressWarnings("static-method")
     public boolean hasArrayElements() {
         return true;
     }